+++ storybook/stories/switch/Switch.template.tsx
@@ -1,0 +1,5 @@
+import { Switch, SwitchProps } from '@equinor/fusion-react-switch';
+
+export const RadioTemplate = (props: SwitchProps): JSX.Element => <Switch {...props} />;
+
+export default RadioTemplate;